requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
braintree==3.29.2
celery==3.1.24
colorama==0.3.7
decorator==4.0.10
Django==1.10
django-crispy-forms==1.6.0
django-filter==0.14.0
django-localflavor==1.3
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-phonenumber-field==1.1.0
django-registration-redux==1.4
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
entrypoints==0.2.2
httpie==0.9.6
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
kombu==3.0.37
markdown2==2.3.1
MarkupSafe==0.23
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.3
phonenumberslite==7.7.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.8
PyJWT==1.4.2
python-dateutil==2.5.3
pytz==2016.6.1
pyzmq==16.0.0
qtconsole==4.2.1
rebar==0.1
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
stripe==1.38.0
tornado==4.4.2
traitlets==4.3.1
virtualenv==15.1.0
wcwidth==0.1.7
widgetsnbextension==1.2.6
win-unicode-console==0.5
